git-learning-game/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
aiofiles==23.2.1
gitpython==3.1.40
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
